+++ src/client/src/commons/form/translationText.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import { withTranslation } from "react-i18next";
 
@@ -7,33 +6,6 @@ class TranslationText extends React.Component {
   render() {
     const { append, extra, id, ns, prepend, t } = this.props;
 
-    if (this.props.developer.debug === true) {
-      return (
-        <span
-          style={{
-            backgroundColor: "#ececec",
-            color: "black",
-            lineHeight: "1.2em",
-            margin: "5px",
-            padding: "5px",
-            ...this.props.style,
-          }}
-          title={`trans=${ns !== "common" ? ns + "." : ""}${id}`}>
-          {prepend}
-          {t(`${ns}:${id}`, extra)}
-          {append}
-          <br />
-          <span
-            style={{
-              color: "red",
-              margin: "5px",
-            }}
-            title={`trans=${ns !== "common" ? ns + "." : ""}${id}`}>
-            trans={`${ns !== "common" ? ns + "." : ""}${id}`}
-          </span>
-        </span>
-      );
-    }
     let text = t(`${ns}:${id}`, extra);
     if (this.props.firstUpperCase === true) {
       return prepend + text.charAt(0).toUpperCase() + text.slice(1) + append;
@@ -42,17 +14,8 @@ class TranslationText extends React.Component {
   }
 }
 
-const mapStateToProps = state => {
-  return {
-    developer: state.developer,
-  };
-};
-
 TranslationText.propTypes = {
   append: PropTypes.string,
-  developer: PropTypes.shape({
-    debug: PropTypes.bool,
-  }),
   extra: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
 
   firstUpperCase: PropTypes.bool,
@@ -71,5 +34,5 @@ TranslationText.defaultProps = {
   prepend: "",
   style: {},
 };
-const ConnectedTranslationText = connect(mapStateToProps, null)(withTranslation(["common"])(TranslationText));
+const ConnectedTranslationText = withTranslation(["common"])(TranslationText);
 export default ConnectedTranslationText;